+++ stac_check/lint.py
@@ -162,6 +162,14 @@ class Linter:
     def check_percent_encoded(self):
         return self.asset_type == "ITEM" and "/" in self.object_id or ":" in self.object_id
 
+    def check_thumbnail(self):
+        if "assets" in self.data:
+            if "thumbnail" in self.data["assets"]:
+                if "type" in self.data["assets"]["thumbnail"]:
+                    if "png" in self.data["assets"]["thumbnail"]["type"] or "jpeg" in self.data["assets"]["thumbnail"]["type"] or \
+                        "jpg" in self.data["assets"]["thumbnail"]["type"] or "webp" in self.data["assets"]["thumbnail"]["type"]:
+                        return True
+
     def create_best_practices_msg(self):
         best_practices = list()
         base_string = "STAC Best Practices: "
@@ -188,7 +196,7 @@ class Linter:
         # best practices - collections should contain summaries
         if self.asset_type == "COLLECTION" and self.summaries == False:
             string_1 = f"    A STAC collection should contain a summaries field"
-            string_2 = f"    https://github.com/radiantearth/stac-spec/blob/master/collection-spec/collection-spec.md"
+            string_2 = f"    It is recommended to store information like eo:bands in summaries"
             best_practices.extend([string_1, string_2, ""])
 
         # best practices - datetime files should not be set to null
@@ -215,6 +223,11 @@ class Linter:
         # best practices - check for bloated metadata in properties
         if self.bloated_metadata:
             string_1 = f"    You have {len(self.data['properties'])} properties. Please consider using links to avoid bloated metadata"
+            best_practices.extend([string_1, ""])
+
+        # best practices - ensure thumbnail is a small file size ["png", "jpeg", "jpg", "webp"]
+        if not self.check_thumbnail() and self.asset_type == "ITEM":
+            string_1 = f"    A thumbnail should have a small file size ie. png, jpeg, jpg, webp"
             best_practices.extend([string_1, ""])
 
         return best_practices